Issue 85 create settings close #85 close #83 close #74 close #73

+++ domain/src/main/java/com/helpfulapps/domain/repository/WeatherRepository.kt
@@ -8,7 +8,7 @@ interface WeatherRepository {
 
     fun downloadForecast(city: String): Completable
 
-    fun downloadForecast(lat: Long, lon: Long): Completable
+    fun downloadForecast(lat: Double, lon: Double): Completable
 
     fun getForecastForAlarms(): Single<List<DayWeather>>
